+++ src/main/java/net/atlas/combatify/mixin/FoodDataMixin.java
@@ -3,9 +3,10 @@ package net.atlas.combatify.mixin;
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.llamalad7.mixinextras.sugar.Local;
 import net.atlas.combatify.Combatify;
-import net.minecraft.util.Mth;
-import net.minecraft.util.RandomSource;
+import net.atlas.combatify.config.HealingMode;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.food.FoodData;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -22,6 +23,12 @@ public class FoodDataMixin {
 
 	@Shadow
 	private float saturationLevel;
+
+	@WrapOperation(method = "add", at = @At(value = "INVOKE", target = "Lnet/minecraft/util/Mth;clamp(FFF)F"))
+	public float capAt20(float val, float min, float max, Operation<Float> original, @Local(ordinal = 0, argsOnly = true) float saturation) {
+		if (Combatify.CONFIG.ctsSaturationCap()) return Math.max(saturationLevel, saturation);
+		return original.call(val, min, Combatify.CONFIG.healingMode() == HealingMode.VANILLA ? max : 20);
+	}
 
 	@ModifyConstant(method = "tick", constant = @Constant(intValue = 18))
 	public int changeConst(int constant) {
@@ -45,12 +52,17 @@ public class FoodDataMixin {
 		return (int) (Combatify.CONFIG.healingTime() * 20);
 	}
 
+	@ModifyConstant(method = "tick", constant = @Constant(intValue = 80,ordinal = 1))
+	public int redirectTickTimer2(int constant) {
+		return (int) (Combatify.CONFIG.starvingTime() * 20);
+	}
+
 	@WrapOperation(method = "tick", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/food/FoodData;addExhaustion(F)V",ordinal = 1))
-	public void modifyNaturalHealing(FoodData instance, float exhaustion, Operation<Void> original) {
+	public void modifyNaturalHealing(FoodData instance, float exhaustion, Operation<Void> original, @Local(argsOnly = true) Player player) {
 		switch (Combatify.CONFIG.healingMode()) {
 			case VANILLA -> original.call(instance, exhaustion);
 			case CTS -> {
-				if (Mth.randomBetweenInclusive(RandomSource.create(), 1, 2) == 2) --foodLevel;
+				if (player.level().random.nextBoolean()) --foodLevel;
 			}
 			case NEW -> {
 				if (saturationLevel > 5.0) original.call(instance, exhaustion);